initial class isolation per #3183
