+++ pycaret/regression/oop.py
@@ -20,8 +20,8 @@ from pycaret.internal.parallel.parallel_backend import ParallelBackend
 # Own module
 from pycaret.internal.pipeline import Pipeline as InternalPipeline
 from pycaret.internal.preprocess.preprocessor import Preprocessor
-from pycaret.internal.pycaret_experiment.supervised_experiment import (
-    _SupervisedExperiment,
+from pycaret.internal.pycaret_experiment.non_ts_supervised_experiment import (
+    _NonTSSupervisedExperiment,
 )
 from pycaret.loggers.base_logger import BaseLogger
 from pycaret.utils.constants import DATAFRAME_LIKE, SEQUENCE_LIKE, TARGET_LIKE
@@ -30,7 +30,7 @@ from pycaret.utils.generic import MLUsecase, highlight_setup
 LOGGER = get_logger()
 
 
-class RegressionExperiment(_SupervisedExperiment, Preprocessor):
+class RegressionExperiment(_NonTSSupervisedExperiment, Preprocessor):
     def __init__(self) -> None:
         super().__init__()
         self._ml_usecase = MLUsecase.REGRESSION
