+++ pycaret/time_series/forecasting/oop.py
@@ -36,7 +36,6 @@ from pycaret.internal.distributions import get_base_distributions
 from pycaret.internal.logging import get_logger, redirect_output
 from pycaret.internal.parallel.parallel_backend import ParallelBackend
 
-# from pycaret.internal.pipeline import get_pipeline_fit_kwargs
 from pycaret.internal.plots.time_series import _get_plot
 from pycaret.internal.plots.utils.time_series import (
     _clean_model_results_labels,
@@ -47,8 +46,8 @@ from pycaret.internal.plots.utils.time_series import (
 from pycaret.internal.preprocess.time_series.forecasting.preprocessor import (
     TSForecastingPreprocessor,
 )
-from pycaret.internal.pycaret_experiment.supervised_experiment import (
-    _SupervisedExperiment,
+from pycaret.internal.pycaret_experiment.ts_supervised_experiment import (
+    _TSSupervisedExperiment,
 )
 from pycaret.internal.tests.time_series import (
     recommend_lowercase_d,
@@ -85,13 +84,13 @@ from pycaret.utils.time_series.forecasting.pipeline import (
     _add_model_to_pipeline,
     _get_imputed_data,
     _get_pipeline_estimator_label,
-    _pipeline_transform,
+    
 )
 
 LOGGER = get_logger()
 
 
-class TSForecastingExperiment(_SupervisedExperiment, TSForecastingPreprocessor):
+class TSForecastingExperiment(_TSSupervisedExperiment, TSForecastingPreprocessor):
     def __init__(self) -> None:
         super().__init__()
         self._ml_usecase = MLUsecase.TIME_SERIES
@@ -5587,159 +5586,7 @@ class TSForecastingExperiment(_SupervisedExperiment, TSForecastingPreprocessor):
         )
         return pipeline_to_use
 
-    @property
-    def X(self):
-        X = self.dataset.drop(self.target_param, axis=1)
-        if X.empty and self.fe_exogenous is None:
-            return None
-        else:
-            # If X is not empty or empty but self.fe_exogenous is provided
-            # Return X instead of None, since the index can be used to
-            # generate features using self.fe_exogenous
-            return X
-
-    @property
-    def dataset_transformed(self):
-        # Use fully trained pipeline to get the requested data
-        return pd.concat(
-            [
-                *_pipeline_transform(
-                    pipeline=self.pipeline_fully_trained, y=self.y, X=self.X
-                )
-            ],
-            axis=1,
-        )
-
-    @property
-    def X_train_transformed(self):
-        # Use pipeline trained on training data only to get the requested data
-        # In time series, the order of arguments and returns may be reversed.
-        return _pipeline_transform(
-            pipeline=self.pipeline, y=self.y_train, X=self.X_train
-        )[1]
-
-    @property
-    def train_transformed(self):
-        # Use pipeline trained on training data only to get the requested data
-        # In time series, the order of arguments and returns may be reversed.
-        return pd.concat(
-            [
-                *_pipeline_transform(
-                    pipeline=self.pipeline, y=self.y_train, X=self.X_train
-                )
-            ],
-            axis=1,
-        )
-
-    @property
-    def X_transformed(self):
-        # Use fully trained pipeline to get the requested data
-        # In time series, the order of arguments and returns may be reversed.
-        return _pipeline_transform(
-            pipeline=self.pipeline_fully_trained, y=self.y, X=self.X
-        )[1]
-
-    @property
-    def X_train(self):
-        X_train = self.train.drop(self.target_param, axis=1)
-
-        if X_train.empty and self.fe_exogenous is None:
-            return None
-        else:
-            # If X_train is not empty or empty but self.fe_exogenous is provided
-            # Return X_train instead of None, since the index can be used to
-            # generate features using self.fe_exogenous
-            return X_train
-
-    @property
-    def X_test(self):
-        # Use index for y_test (idx 2) to get the data
-        test = self.dataset.loc[self.idx[2], :]
-        X_test = test.drop(self.target_param, axis=1)
-
-        if X_test.empty and self.fe_exogenous is None:
-            return None
-        else:
-            # If X_test is not empty or empty but self.fe_exogenous is provided
-            # Return X_test instead of None, since the index can be used to
-            # generate features using self.fe_exogenous
-            return X_test
-
-    @property
-    def test(self):
-        # Return the y_test indices not X_test indices.
-        # X_test indices are expanded indices for handling FH with gaps.
-        # But if we return X_test indices, then we will get expanded test
-        # indices even for univariate time series without exogenous variables
-        # which would be confusing. Hence, we return y_test indices here and if
-        # we want to get X_test indices, then we use self.X_test directly.
-        # Refer:
-        # https://github.com/sktime/sktime/issues/2598#issuecomment-1203308542
-        # https://github.com/sktime/sktime/blob/4164639e1c521b112711c045d0f7e63013c1e4eb/sktime/forecasting/model_evaluation/_functions.py#L196
-        return self.dataset.loc[self.idx[1], :]
-
-    @property
-    def test_transformed(self):
-        # When transforming the test set, we can and should use all data before that
-        # In time series, the order of arguments and returns may be reversed.
-        all_data = pd.concat(
-            [
-                *_pipeline_transform(
-                    pipeline=self.pipeline_fully_trained,
-                    y=self.y,
-                    X=self.X,
-                )
-            ],
-            axis=1,
-        )
-        # Return the y_test indices not X_test indices.
-        # X_test indices are expanded indices for handling FH with gaps.
-        # But if we return X_test indices, then we will get expanded test
-        # indices even for univariate time series without exogenous variables
-        # which would be confusing. Hence, we return y_test indices here and if
-        # we want to get X_test indices, then we use self.X_test directly.
-        # Refer:
-        # https://github.com/sktime/sktime/issues/2598#issuecomment-1203308542
-        # https://github.com/sktime/sktime/blob/4164639e1c521b112711c045d0f7e63013c1e4eb/sktime/forecasting/model_evaluation/_functions.py#L196
-        return all_data.loc[self.idx[1]]
-
-    @property
-    def y_transformed(self):
-        # Use fully trained pipeline to get the requested data
-        # In time series, the order of arguments and returns may be reversed.
-        return _pipeline_transform(
-            pipeline=self.pipeline_fully_trained, y=self.y, X=self.X
-        )[0]
-
-    @property
-    def X_test_transformed(self):
-        # In time series, the order of arguments and returns may be reversed.
-        # When transforming the test set, we can and should use all data before that
-        _, X = _pipeline_transform(
-            pipeline=self.pipeline_fully_trained, y=self.y, X=self.X
-        )
-
-        if X is None:
-            return None
-        else:
-            return X.loc[self.idx[2]]
-
-    @property
-    def y_train_transformed(self):
-        # Use pipeline trained on training data only to get the requested data
-        # In time series, the order of arguments and returns may be reversed.
-        return _pipeline_transform(
-            pipeline=self.pipeline, y=self.y_train, X=self.X_train
-        )[0]
-
-    @property
-    def y_test_transformed(self):
-        # In time series, the order of arguments and returns may be reversed.
-        # When transforming the test set, we can and should use all data before that
-        y, _ = _pipeline_transform(
-            pipeline=self.pipeline_fully_trained, y=self.y, X=self.X
-        )
-        return y.loc[self.idx[1]]
+    
 
 
 def _validate_split_requested(split: str):
